Make CacheMap main solution
Move all others into internal folder
Make CacheMap work without losing performance because of the moving
out of the internal modules.
