+++ benchmark/CacheMap.hs
@@ -25,7 +25,7 @@ import Data.Proxy (Proxy (..))
 import Data.Typeable (Typeable)
 import GHC.TypeLits
 
-import Data.TypeRep.CacheMap (TF (..), TypeRepMap (..), fromList, lookup)
+import Data.TypeRep.Map (TF (..), TypeRepMap (..), fromList, lookup)
 
 benchCacheMap :: Benchmark
 benchCacheMap = bgroup "vector optimal cache"
