+++ benchmark/OptimalVector.hs
@@ -25,7 +25,7 @@ import Data.Proxy (Proxy (..))
 import Data.Typeable (Typeable)
 import GHC.TypeLits
 
-import Data.TypeRep.Map (TF (..), TypeRepMap (..), fromList, lookup)
+import Data.TypeRep.OptimalVector (TF (..), TypeRepMap (..), fromList, lookup)
 
 benchVectorOpt :: Benchmark
 benchVectorOpt = bgroup "vector optimal"
