+++ stack.yaml
@@ -1,1 +1,1 @@
-resolver: lts-11.15
+resolver: nightly-2018-07-05
