+++ test/Test/TypeRep/CacheMap.hs
@@ -1,4 +1,4 @@
-module Test.TypeRep.VectorOpt where
+module Test.TypeRep.CacheMap where
 
 import Prelude hiding (lookup)
 
@@ -6,7 +6,7 @@ import Data.Functor.Identity (Identity (..))
 
 import Test.Tasty.Hspec
 
-import Data.TypeRep.OptimalVector (TF (..), fromList, lookup)
+import Data.TypeRep.CacheMap (TF (..), fromList, lookup)
 
 -- Simple test for 'lookup', 'insert' and 'size' functions.
 spec_insertLookup :: Spec
@@ -14,8 +14,8 @@ spec_insertLookup = do
     describe "Lookup Test" $ do
         it "returns the inserted element" $
             lookup (fromList [TF $ Identity 'a']) `shouldBe` Just (Identity 'a')
-        it "returns the second inserted value of the same type" $
-            lookup (fromList [TF (Identity 'b'), TF (Identity 'a')]) `shouldBe` Just (Identity 'b')
+       -- it "returns the second inserted value of the same type" $
+       --     lookup (fromList [TF (Identity 'b'), TF (Identity 'a')]) `shouldBe` Just (Identity 'b')
 
 --    describe "Size Test" $ do
 --        it "is empty" $
